0.3.0-beta

+++ src/controllers/videoController.ts
@@ -2,20 +2,23 @@
 import express from "express";
 import { MulterError } from "multer";
 import { getVideoDurationInSeconds } from "get-video-duration";
-
-import { MulterFileType, VideoType, PeekoRequest } from "../types";
+import SimpleThumbnail from "simple-thumbnail-ts";
+import ffmpegPath from "ffmpeg-static";
 
 import { checkVideoExists } from "../middleware/checkResourceExists";
 import { requireAuth } from "../middleware/authentication";
 
-import { upload, deleteFile } from "../models/multerSetup";
+import { upload, deleteFile } from "../middleware/multer";
 import Video from "../models/video";
 import Comment from "../models/comment";
 import User from "../models/user";
-import { s3_get, s3_post, s3_delete } from "../models/s3";
+import { s3_get, s3_post, s3_delete } from "../utils/s3";
+import { PipelineStage } from "mongoose";
+import { replaceArrayWithCount } from "../utils/utils";
 
 // create router
 const router = express.Router();
+export const thumbnailSuffix = "-thumbnail.png";
 
 /**
  * @post
@@ -35,7 +38,6 @@ router.post("/uploadVideo", requireAuth, async (req: PeekoRequest, res) => {
             });
         }
 
-        // if no errors
         // validate file existence
         if (!req.file) {
             return res.status(400).json({
@@ -45,77 +47,113 @@ router.post("/uploadVideo", requireAuth, async (req: PeekoRequest, res) => {
         }
 
         // destructure
-        const uploaderId = req.currentUser!._id;
-        const uploaderUsername = req.currentUser!.username;
         const videoFile: MulterFileType = req.file;
+
+        const thumbnailObject = {
+            path: videoFile.path + thumbnailSuffix,
+            filename: videoFile.filename + thumbnailSuffix,
+            mimetype: "image/png",
+        };
 
         try {
             // validate video file duration
             const duration = await getVideoDurationInSeconds(videoFile.path);
             if (duration > 300) {
-                await deleteFile(videoFile.path);
                 return res.status(400).json({
                     success: false,
                     error: "Video duration must not exceed 5 minutes",
                 });
             }
 
+            // create thumbnail
+            await new SimpleThumbnail().generate(
+                videoFile.path,
+                thumbnailObject.path,
+                "480x854",
+                {
+                    path: ffmpegPath,
+                }
+            );
+
             // upload to S3
-            const videoKey = await s3_post(videoFile);
-            await deleteFile(videoFile.path);
+            const [videoKey] = await Promise.all([
+                s3_post(videoFile),
+                s3_post(thumbnailObject),
+            ]);
 
             // create video object structure
             const videoObject = {
-                uploaderId,
-                uploaderUsername,
+                uploader: req.currentUser!._id,
                 videoKey,
             };
 
             // upload video data to db
-            const videoDocument: VideoType = await Video.create(videoObject);
+            let videoDocument: VideoType = await Video.create(videoObject);
+            videoDocument = await videoDocument.populate("uploader", {
+                username: 1,
+            });
 
             // return result to client
             res.status(200).json({
                 success: true,
-                videoDocument,
+                videoDocument: replaceArrayWithCount(videoDocument),
             });
         } catch (err: any) {
             console.error(err);
+
+            // delete video and thumbnail if they were uploaded to s3
+            await s3_delete([
+                videoFile.filename,
+                videoFile.filename + thumbnailSuffix,
+            ]);
+
             return res.status(400).json({
                 success: false,
                 error: err.message,
             });
+        } finally {
+            await deleteFile([videoFile.path, thumbnailObject.path]);
         }
     });
 });
 
 /**
  * @get
- *      GET request to get a specific video file through a provided video key
+ *      GET request to get a specific video thumbnail through a proided video key
  */
-router.get("/streamVideo/:videoKey", checkVideoExists, async (req, res) => {
-    // destructure
-    const { videoKey } = req.params;
+router.get(
+    ["/streamThumbnail/:videoKey", "/streamVideo/:videoKey"],
+    checkVideoExists,
+    async (req, res) => {
+        const { videoKey } = req.params;
+        const isThumbnail = req.path.includes("streamThumbnail");
 
-    try {
-        // get video from s3 and stream it back to client
-        const result = await s3_get(videoKey);
-        const stream = result.Body as NodeJS.ReadableStream;
+        try {
+            // get requested resource from s3 and stream it back to client
+            // if requesting thumbnail, add thumbnail key suffix
+            const result = await s3_get(
+                videoKey + (isThumbnail ? thumbnailSuffix : "")
+            );
+            const stream = result.Body as NodeJS.ReadableStream;
 
-        res.header({
-            "content-type": result.ContentType,
-            "content-length": result.ContentLength,
-        });
+            // if requesting video, set "content-length"
+            res.header({
+                "content-type": result.ContentType,
+                ...(isThumbnail
+                    ? {}
+                    : { "content-length": result.ContentLength }),
+            });
 
-        stream.pipe(res);
-    } catch (err: any) {
-        console.error(err);
-        res.status(400).json({
-            success: false,
-            error: err.message,
-        });
+            stream.pipe(res);
+        } catch (err: any) {
+            console.error(err);
+            res.status(400).json({
+                success: false,
+                error: err.message,
+            });
+        }
     }
-});
+);
 
 /**
  * @get
@@ -127,18 +165,23 @@ router.get(
     async (req: PeekoRequest, res) => {
         try {
             // get video from db
-            const videoDocument = req.resource as VideoType;
+            let videoDocument = req.resource as VideoType;
 
-            // set video as viewed
-            await User.findByIdAndUpdate(req.currentUser?._id, {
-                $addToSet: {
-                    viewed: videoDocument.videoKey,
-                },
-            });
+            // if user signed in, register view
+            if (req.currentUser) {
+                await Video.findByIdAndUpdate(videoDocument._id, {
+                    $push: { views: req.currentUser._id },
+                });
+            }
+
+            const isLiked = !!videoDocument.likes?.includes(
+                req.currentUser?._id
+            );
 
             res.status(200).json({
                 success: true,
-                videoDocument,
+                videoDocument: replaceArrayWithCount(videoDocument),
+                isLiked,
             });
         } catch (err: any) {
             console.error(err);
@@ -157,40 +200,69 @@ router.get(
 router.get("/getVideos/:count?", async (req: PeekoRequest, res) => {
     // destructure
     const count = parseInt(req.params.count as string) || 10;
-    const viewed: string[] = req.currentUser?.viewed || [];
 
     try {
         /**
-         * Use MongoDB $match aggregation stage to exclude
+         * Use $match stage to exclude
          * previously watched videos from the pipeline.
          *
-         * Use MongoDB $sample aggregation stage to get a
+         * Use $sample stage to get a
          * #count number of randomly selected videos.
          *
-         * Use MongoDB $group aggregation stage to make sure
+         * Use $lookup stage to populate uploader field
+         *
+         * Use $project stage to define what fields to include
+         *
+         * Use $group & $replaceWith stages to make sure
          * we don't get duplicate documents.
          */
-        let videoDocuments: VideoType[] = await Video.aggregate([
-            { $match: { videoKey: { $nin: viewed } } },
+
+        let matchStage = req.currentUser
+            ? { views: { $ne: req.currentUser._id } }
+            : {};
+
+        const aggregationPipeline: PipelineStage[] = [
+            { $match: matchStage },
             { $sample: { size: count } },
+            {
+                $lookup: {
+                    from: User.collection.name,
+                    localField: "uploader",
+                    foreignField: "_id",
+                    as: "uploader",
+                },
+            },
+            { $unwind: "$uploader" },
+            {
+                $project: {
+                    uploader: {
+                        _id: 1,
+                        username: 1,
+                    },
+                    videoKey: 1,
+                    likes: 1,
+                    views: 1,
+                    commentsCount: 1,
+                    createdAt: 1,
+                    updatedAt: 1,
+                },
+            },
             { $group: { _id: "$_id", doc: { $first: "$$ROOT" } } },
             { $replaceWith: "$doc" },
-        ]);
+        ];
 
-        // backup in case no unseen videos found
+        let videoDocuments = await Video.aggregate(aggregationPipeline);
+
+        // backup in case no unveiwed videos found
         if (videoDocuments.length === 0) {
-            videoDocuments = await Video.aggregate([
-                { $match: {} },
-                { $sample: { size: count } },
-                { $group: { _id: "$_id", doc: { $first: "$$ROOT" } } },
-                { $replaceWith: "$doc" },
-            ]);
+            matchStage = {};
+            videoDocuments = await Video.aggregate(aggregationPipeline);
         }
 
         // return result
         return res.status(200).json({
             success: true,
-            videoDocuments,
+            videoDocuments: replaceArrayWithCount(videoDocuments),
         });
     } catch (err: any) {
         console.error(err);
@@ -216,7 +288,10 @@ router.delete(
 
         try {
             // if deleter is not publisher
-            if (req.currentUser!._id.toString() !== videoDocument.uploaderId) {
+            if (
+                req.currentUser!._id.toString() !==
+                videoDocument.uploader._id.toString()
+            ) {
                 return res.status(400).json({
                     success: false,
                     error: "Unauthorized Action",
@@ -224,11 +299,27 @@ router.delete(
             }
 
             // delete video document from db
-            const deletedVideoDocument: VideoType | null =
-                await Video.findOneAndDelete({ videoKey });
+            let deletedVideoDocument: VideoType | null =
+                await Video.findOneAndDelete({
+                    videoKey,
+                });
 
-            // delete video file from s3
-            await s3_delete(videoKey);
+            if (!deletedVideoDocument) {
+                return res.status(400).json({
+                    success: false,
+                    error: "Video not found",
+                });
+            }
+
+            deletedVideoDocument = await deletedVideoDocument.populate(
+                "uploader",
+                {
+                    username: 1,
+                }
+            );
+
+            // delete video file & thumbnail from s3
+            await s3_delete([videoKey, videoKey + thumbnailSuffix]);
 
             // delete comments associated with video
             await Comment.deleteMany({ videoKey });
@@ -236,7 +327,8 @@ router.delete(
             // return response
             res.status(200).json({
                 success: true,
-                deletedVideoDocument,
+                deletedVideoDocument:
+                    replaceArrayWithCount(deletedVideoDocument),
             });
         } catch (err: any) {
             console.error(err);
